refactor: replace AnimatedMovieCard with MovieCard component for improved structure and functionality
