+++ src/features/movies/components/MovieCard/index.ts
@@ -1,2 +1,2 @@
-export { default } from './MovieCard';
-export type { MovieCardProps } from './MovieCard.types';
+export { default } from "./MovieCard";
+export type { MovieCardProps } from "./MovieCard.types";
